sleazy solution to doubling problem
I'm not why and how we're getting multiple subscribed, but... I'm really not getting paid enough to figure that out.  This works.

All we do is every time we append a message to the message channel frame, we delete messages from the DOM with the same ID.

Which is sort of sleazy, but it works, and it's sort of a blunt instrument against anything else that can cause the same problem.